+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Added
 - log files that match particular formats with `-log fmt/1,@set2` (comma separated list of format IDs/format sets). These can be mixed with regular log options e.g. `-log unknown,fmt/1,chart`
 - generate a summary view of formats matched during a scan with `-log chart` (or just `-log c`)
-- replay scans from results files with `sf -r`: load one or more results files to replay logging or to convert to a different output format e.g. `sf -r -csv results.yaml` or `sf -r -log unknown,chart,stdout results1.yaml results2.csv`
+- replay scans from results files with `sf -replay`: load one or more results files to replay logging or to convert to a different output format e.g. `sf -replay -csv results.yaml` or `sf -replay -log unknown,chart,stdout results1.yaml results2.csv`
 - compare results with `roy compare` subcommand: view the difference between two or more results e.g. `roy compare results1.yaml results2.csv droid.csv ...`
 - `roy sets` subcommand: `roy sets` creates pronom-all.json, pronom-families.json, and pronom-types.json sets files;
 `roy sets -changes` creates a pronom-changes.json sets file from a PRONOM release-notes.xml file; `roy sets -list @set1,@set2` lists contents of a comma-separated list of format sets